fix: handle missing endpoints for azureb2c
Previously, if the metadata was missing an endpoint definition, we would get a null pointer exception when trying to call `toASCIIString()` on the null uri. Enhanced the logic to handle this case and added `missingEndpoints` as part of the `MetadataValidationResult` to be used when providing error message details.

risk: low